Day 9 prework

+++ src/main/java/dev/haan/aoc2019/intcode/Reader.java
@@ -2,5 +2,5 @@ package dev.haan.aoc2019.intcode;
 
 public interface Reader {
 
-    int read() throws InterruptedException;
+    long read() throws InterruptedException;
 }
