+++ src/main/java/dev/haan/aoc2019/intcode/Writer.java
@@ -2,5 +2,5 @@ package dev.haan.aoc2019.intcode;
 
 public interface Writer {
 
-    void write(int value) throws InterruptedException;
+    void write(long value) throws InterruptedException;
 }
